// TASK_MANAGEMENT_REFACTORED/frontend/src/store/auth/types.ts
export enum AUTH_ACTION_TYPES {
	SET_ACCESS_TOKEN = "auth/SET_ACCESS_TOKEN",
	REMOVE_ACCESS_TOKEN = "auth/REMOVE_ACCESS_TOKEN",
	SET_REFRESH_TOKEN = "auth/SET_REFRESH_TOKEN",
	REMOVE_REFRESH_TOKEN = "auth/REMOVE_REFRESH_TOKEN",

	LOGIN_START = "auth/LOGIN_START",
	LOGIN_FAILURE = "auth/LOGIN_FAILURE",
	FETCH_REMOTE_USER_START = "auth/FETCH_REMOTE_USER_START",
	FETCH_REMOTE_USER_FAILURE = "auth/FETCH_REMOTE_USER_FAILURE",
	SET_USER = "auth/SET_USER",

	FETCH_UP_TO_DATE_INSTITUTION = "auth/FETCH_UP_TO_DATE_INSTITUTION",

	LOGOUT_START = "auth/LOGOUT_START",
	LOGOUT_FAILURE = "auth/LOGOUT_FAILURE",
	LOGOUT_SUCCESS = "auth/LOGOUT_SUCCESS",
	CLEAR_AUTH_ERROR = "auth/CLEAR_AUTH_ERROR",

	SET_SELECTED_INSTITUTION = "auth/SET_SELECTED_INSTITUTION",
	ADD_NEW_Institution = "auth/ADD_NEW_Institution",
	SET_ATTACHED_INSTITUTIONS = "auth/SET_ATTACHED_INSTITUTIONS",
	SET_SELECTED_BRANCH = "auth/SET_SELECTED_BRANCH",

	UPDATE_THEME = "auth/UPDATE_THEME",
	REMOVE_THEME = "auth/REMOVE_THEME",

	SET_TEMPORARY_PERMISSIONS = "auth/SET_TEMPORARY_PERMISSIONS",
	CLEAR_TEMPORARY_PERMISSIONS = "auth/CLEAR_TEMPORARY_PERMISSIONS",

	SET_SELECTED_TILL = "auth/SET_SELECTED_TILL",
	CLEAR_SELECTED_TILL = "auth/CLEAR_SELECTED_TILL",

	// Auto Logout mechanism

	USER_ACTIVITY_DETECTED = "auth/USER_ACTIVITY_DETECTED",
	SHOW_LOGOUT_WARNING = "auth/SHOW_LOGOUT_WARNING",
	HIDE_LOGOUT_WARNING = "auth/HIDE_LOGOUT_WARNING",
	CONFIRM_LOGOUT = "auth/CONFIRM_LOGOUT",
	CANCEL_LOGOUT = "auth/CANCEL_LOGOUT",
	REFRESH_TOKENS_START = "auth/REFRESH_TOKENS_START",
	REFRESH_TOKENS_SUCCESS = "auth/REFRESH_TOKENS_SUCCESS",
	REFRESH_TOKENS_FAILURE = "auth/REFRESH_TOKENS_FAILURE",
	SET_INACTIVITY_TIMEOUT = "auth/SET_INACTIVITY_TIMEOUT",
}